requirements.txt
websockets==11.0.3
python-dotenv==1.0.0